# docs/swagger.yaml
definitions:
  api.JSONNumReadResult:
    properties:
      stat_name1:
        type: number
      stat_name2:
        type: number
    type: object
  api.NumStatRegistration:
    properties:
      label:
        type: string
      name:
        type: string
      value:
        type: number
    type: object
  api.ValueResponse:
    type: object
info:
  contact: {}
paths:
  /read/num/exponentialrate/{name}:
    get:
      consumes:
      - application/json
      description: Read rate estimates of a specific statistic. Not suitable for accuracy,
        but great for quickly comparing stats.
      parameters:
      - description: Stat name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ValueResponse'
      summary: Read all rate estimates for a stat name
  /read/num/exponentialrate/{name}/{label}:
    get:
      consumes:
      - application/json
      description: Read rate estimate of a specific statistic. Not suitable for accuracy,
        but great for quickly comparing stats.
      parameters:
      - description: Stat name
        in: path
        name: name
        required: true
        type: string
      - description: Stat label
        in: path
        name: label
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ValueResponse'
      summary: Read an exponential *estimate* of a numeric rate of change
  /read/num/mean/{name}:
    get:
      consumes:
      - application/json
      description: Returns the mean for all values across the stat name
      parameters:
      - description: Stat name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ValueResponse'
      summary: Read the mean value for a name
  /read/num/names:
    get:
      consumes:
      - application/json
      description: Read list of numeric stat names
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.JSONNumReadResult'
      summary: Read all stat names
  /read/num/value/{name}:
    get:
      consumes:
      - application/json
      description: Read all labels associated with a stat by name
      parameters:
      - description: Stat name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.JSONNumReadResult'
      summary: Read a specific statistic
  /read/num/value/{name}/{label}:
    get:
      consumes:
      - application/json
      description: Read a specific single numeric statistic by name and label
      parameters:
      - description: Stat name
        in: path
        name: name
        required: true
        type: string
      - description: Stat label
        in: path
        name: label
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ValueResponse'
      summary: Read a specific statistic
  /register/num/decrease:
    post:
      consumes:
      - application/json
      description: Decrease a numeric stat by some amount
      parameters:
      - description: Stat and value to decrease by
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/api.NumStatRegistration'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ValueResponse'
      summary: Decrease a numeric stat
  /register/num/increase:
    post:
      consumes:
      - application/json
      description: Increase a numeric stat by some amount
      parameters:
      - description: Stat and value
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/api.NumStatRegistration'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ValueResponse'
      summary: Increase a numeric stat
  /register/num/set:
    post:
      consumes:
      - application/json
      description: Set a numeric stat
      parameters:
      - description: Stat and value to increase by
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/api.NumStatRegistration'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ValueResponse'
      summary: Set numeric stat
  /serialize:
    get:
      consumes:
      - application/json
      description: Read list of numeric stat names
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.JSONNumReadResult'
      summary: Read all stat names
swagger: "2.0"